+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/TrackTablePanel.java
@@ -91,7 +91,7 @@ public final class TrackTablePanel extends EntityTablePanel {
 						.control(Control.builder()
 										.command(this::raisePriceOfSelected)
 										.caption(BUNDLE.getString("raise_price") + "...")
-										.enabled(model().selection().empty().not()))
+										.enabled(model().selection().present()))
 						// and a separator
 						.separator()
 						// and add all the default controls